Update to use Bukkit 1.7.10-R0.1 API, rip out old SpoutPlugin support

+++ src/main/java/org/dynmap/bukkit/BukkitVersionHelper.java
@@ -142,5 +142,5 @@ public abstract class BukkitVersionHelper {
     /**
      * Get player health
      */
-    public abstract int getHealth(Player p);
+    public abstract double getHealth(Player p);
 }
